generalize CommunityFilter, relocate CommunityFilter
