+++ src/screens/community/CommunityScreen.js
@@ -2,7 +2,7 @@ import { Text, View, TouchableOpacity, ScrollView } from "react-native";
 import { useState } from "react";
 import CommunityCardLarge from "../../components/community-card/community-card-large/CommunityCardLarge";
 import CommunityCardSmall from "../../components/community-card/community-card-small/CommunityCardSmall";
-import CommunityFilter from "../../components/community-card/community-filter/CommunityFilter";
+import CommunityFilter from "../../components/community-filter/CommunityFilter";
 import styles from "./styles";
 
 const CommunityScreen = () => {
@@ -83,6 +83,63 @@ const CommunityScreen = () => {
       photo: "",
     },
   ]);
+  const [campus, setCampus] = useState([
+    {
+      id: 1,
+      name: "SGS campus",
+    },
+    {
+      id: 2,
+      name: "Hanoi campus",
+    },
+    {
+      id: 3,
+      name: "Danang campus",
+    },
+  ]);
+
+  const [topic, setTopic] = useState([
+    {
+      id: 1,
+      name: "Economics",
+    },
+    {
+      id: 2,
+      name: "Business",
+    },
+    {
+      id: 3,
+      name: "Robotics",
+    },
+    {
+      id: 4,
+      name: "Management",
+    },
+    {
+      id: 5,
+      name: "Aviation",
+    },
+    {
+      id: 6,
+      name: "Education",
+    },
+    {
+      id: 7,
+      name: "Design",
+    },
+    {
+      id: 8,
+      name: "Professional Communication",
+    },
+    {
+      id: 9,
+      name: "Technology",
+    },
+    {
+      id: 10,
+      name: "Digital Marketing",
+    },
+  ]);
 
   const [show, setShow] = useState();
 
@@ -114,7 +171,12 @@ const CommunityScreen = () => {
         >
           <Text style={{ color: "#ffffff" }}>Filter</Text>
         </TouchableOpacity>
-        {show ? <CommunityFilter /> : null}
+        {show ? (
+          <View>
+            <CommunityFilter name={"Campus"} data={campus} />
+            <CommunityFilter name={"Topic"} data={topic} />
+          </View>
+        ) : null}
         <View style={styles.communityList}>
           {others.map((user) => {
             return <CommunityCardSmall id={user.id} name={user.name} />;
